Change vga16 file format
Not fully planar, line oriented: Full width of blue pixels, followed by
full with of red pixels, etc.

This is better than one byte of red, one byte of blue, and so on. Way
less OUTs are necessary to select the destination memory plane..

But still suboptimal for images spawning the full width of the screen...